+++ lib/model_editors.dart
@@ -1,11 +1,18 @@
-export 'src/controllers/abstract_list.dart';
 export 'src/controllers/checkbox_group.dart';
+export 'src/controllers/editor.dart';
+export 'src/controllers/list.dart';
+export 'src/controllers/map.dart';
 export 'src/controllers/with_id_title.dart';
 export 'src/controllers/with_text.dart';
-
 export 'src/widgets/capsule_list_editor.dart';
+export 'src/widgets/collection_add_button_builder.dart';
 export 'src/widgets/column_list_editor.dart';
-export 'src/widgets/list_add_button_builder.dart';
+export 'src/widgets/default_add_button.dart';
+export 'src/widgets/enum_dropdown_button.dart';
 export 'src/widgets/material_checkbox_column.dart';
-export 'src/widgets/reorderable_list_view_editor.dart';
+export 'src/widgets/model_view_or_edit.dart';
+export 'src/widgets/model_view_or_edit_raw.dart';
+export 'src/widgets/reorderable_collection_editor.dart';
+export 'src/widgets/reorderable_list_editor.dart';
+export 'src/widgets/reorderable_map_editor.dart';
 export 'src/widgets/wrap_list_editor.dart';