+++ lib/src/controllers/with_id_title.dart
@@ -2,8 +2,11 @@ import 'package:model_interfaces/model_interfaces.dart';
 
 import 'with_text.dart';
 
-class WithIdTitleEditingController<I, T extends WithIdTitle<I>>
-    extends WithTextEditingController<T> {
+class WithIdTitleEditingController<
+    I,
+    T extends WithIdTitle<I>
+//
+    > extends WithTextEditingController<T> {
   @override
   set value(T? value) {
     textEditingController.text = value?.title ?? '';